web/requirements.txt
asgiref==3.7.2
Django==4.2.5
django-filter==23.3
djangorestframework==3.14.0
psycopg2-binary==2.9.8
python-decouple==3.8
pytz==2023.3.post1
sqlparse==0.4.4
